[feat] Pointer type view on terms: Terms
Added field to StackRef, StaticRef and HeapRef
so we can distinguish if the pointer to this Ref
was casted.
e.g. we have byte[], we get pointer to it byte*
and cast it to int*. We should somewhere store
information about the pointer type which we use
to view this Ref (here: int*, so we store int).
This is what we call `Pointer type view`.